# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/mailing_traces:
    get:
      tags:
        - MailingTraces
      summary: List MailingTraces
      description: Get a paginated list of MailingTraces with optional filtering and sorting
      operationId: listMailingTraces
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingTraceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MailingTraces
      summary: Create MailingTrace
      description: Create a new MailingTrace
      operationId: createMailingTrace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMailingTraceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingTrace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    MailingTrace:
      type: object
      required:
        - id
        - trace_type
        - is_test_trace
        - mailing_id
        - recipient_model
        - recipient_id
        - recipient_email
        - trace_status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        trace_type:
          type: string
        is_test_trace:
          type: boolean
        mailing_id:
          type: string
          format: uuid
        campaign_id:
          nullable: true
          type: string
          format: uuid
        recipient_model:
          type: string
          maxLength: 64
        recipient_id:
          type: string
          format: uuid
        recipient_email:
          type: string
          maxLength: 255
        mail_id:
          nullable: true
          type: string
          format: uuid
        message_id:
          nullable: true
          type: string
          maxLength: 255
        trace_status:
          type: string
        failure_type:
          nullable: true
          type: string
        failure_reason:
          nullable: true
          type: string
        sent_datetime:
          nullable: true
          type: string
          format: date-time
        open_datetime:
          nullable: true
          type: string
          format: date-time
        reply_datetime:
          nullable: true
          type: string
          format: date-time
        links_click_datetime:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateMailingTraceRequest:
      type: object
      required:
        - trace_type
        - is_test_trace
        - mailing_id
        - recipient_model
        - recipient_id
        - recipient_email
        - trace_status
        - metadata
      properties:
        trace_type:
          type: string
        is_test_trace:
          type: boolean
        mailing_id:
          type: string
          format: uuid
        campaign_id:
          nullable: true
          type: string
          format: uuid
        recipient_model:
          type: string
          maxLength: 64
        recipient_id:
          type: string
          format: uuid
        recipient_email:
          type: string
          maxLength: 255
        mail_id:
          nullable: true
          type: string
          format: uuid
        message_id:
          nullable: true
          type: string
          maxLength: 255
        trace_status:
          type: string
        failure_type:
          nullable: true
          type: string
        failure_reason:
          nullable: true
          type: string
        sent_datetime:
          nullable: true
          type: string
          format: date-time
        open_datetime:
          nullable: true
          type: string
          format: date-time
        reply_datetime:
          nullable: true
          type: string
          format: date-time
        links_click_datetime:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    MailingTraceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MailingTrace'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []